Set dest size if not scaling
This resets destination width and height to source width and height
unless the image is going to be scaled. It results images to be centered
on a canvas (when specified) instead of staying at top-left corner when
no scaling operation is in place.

Now we are able to specify both destination and canvas to the same size
and if the source is larger, it's scaled down; if the source is smaller
it's placed in the center of the canvas.